requirements.txt
Pillow==9.3.0
selenium==4.7.2
webdriver-manager==3.8.5
